requirements.txt
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15
PyYAML==5.3.1
Flask==1.1.1
pymysql==0.9.3
boto3==1.13.20
scikit-learn==0.23.1
numpy==1.17.2
pandas==1.0.3
pytest==5.4.1
